pyproject.toml
[project]
name = "integration-platform"
version = "0.1.0"
description = "Generic multi-tenant integration platform for SaaS applications"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "apscheduler>=3.10.4",
    "httpx>=0.25.2",
    "tenacity>=8.2.3",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "ruff>=0.1.6",
    "black>=23.11.0",
    "types-python-jose>=3.3.4",
]
aws = [
    "boto3>=1.34.0",
    "aioboto3>=12.0.0",
]
azure = [
    "azure-storage-blob>=12.19.0",
    "azure-storage-queue>=12.8.0",
    "azure-keyvault-secrets>=4.7.0",
    "azure-identity>=1.15.0",
]
gcp = [
    "google-cloud-storage>=2.14.0",
    "google-cloud-pubsub>=2.18.0",
    "google-cloud-kms>=2.19.0",
]
cloud = [
    "boto3>=1.34.0",
    "aioboto3>=12.0.0",
    "azure-storage-blob>=12.19.0",
    "azure-keyvault-secrets>=4.7.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-kms>=2.19.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=48",  # Unit tests only; integration tests would increase to 65%+
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "apscheduler.*",
    "boto3.*",
    "botocore.*",
    "aioboto3.*",
    "azure.*",
    "google.cloud.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]
